# impala-bridge/openapi.yaml
openapi: 3.1.0
info:
  title: impala-bridge
  version: "0.0.0"
  summary: REST bridge between offline Payala payments and the Stellar network.
  description: |
    impala-bridge issues JWTs, persists dual-chain transactions, dispatches
    notifications via SMS/email/push/webhook, and streams events from Stellar
    Horizon and Payala.

    Authentication uses a two-token strategy:

    - **Refresh token** (14-day TTL) — obtained via username+password from `POST /token`.
      Used only to mint temporal tokens, never for API access.
    - **Temporal token** (1-hour TTL) — obtained from a refresh token via `POST /token`.
      Sent as `Authorization: Bearer <token>` on every authenticated request.

    All mutating endpoints enforce account ownership via `require_owner`. A
    `POST /logout` revokes the current token (the JTI is added to a Redis
    blacklist until it would have expired naturally).

    Rate limits apply to `/authenticate`, `/token`, `/mfa/verify`, `/transaction`,
    and `/card`: 10 requests per 60-second window, keyed by account id (and
    MFA type where applicable). Exceeding the limit returns HTTP 429 with a
    `Retry-After` header.

    Global request timeout is 30 s (HTTP 408 on exceed). Postgres has a
    per-connection `statement_timeout` of 60 s.
  contact:
    name: impala-bridge maintainers
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
...
  - url: https://{host}
...
security:
  - bearerAuth: []
paths:
...
  /admin/exchange-reserve/replenishment:
    get:
      tags: [reserve]
      summary: Replenishment policies and recent cycles (admin, treasurer, or auditor).
      responses:
        '200': { description: Policies plus cycle history. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/exchange-reserve/replenishment/policies/{kind}:
    put:
      tags: [reserve]
      summary: Set a replenishment policy (admin or treasurer).
      description: |
        `max_spend_minor` and `daily_spend_cap_minor` of 0 mean
        *unconfigured, refuse to run* — never "unlimited" — so enabling a
        kind whose caps are still 0 is rejected rather than producing a
        policy that looks armed and silently never runs. `min_float_minor`
        must cover the Stellar base reserve plus fees, or customer payouts
        begin failing. `min_price_minor` and `max_slippage_bps` are the
        guards that stop a mispriced quote dumping the float.

        `usdc_to_usd` cannot be enabled (400): no treasury beneficiary
        configuration exists yet, so every cycle of that kind would abort at
        creation and churn a hold/release journal pair per cooldown. It can
        still be updated with `enabled: false`.
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [xlm_to_usdc, usdc_to_usd] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplenishPolicyUpdateRequest' }
      responses:
        '200': { description: Policy updated. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/exchange-reserve/replenishment/run:
    post:
      tags: [reserve]
      summary: Start one replenishment cycle now (admin or treasurer; rate limited 5/60s).
      description: |
        Takes the reserve watcher's advisory lock and returns 409 if the
        watcher is mid-tick — the reserve account signs from a single
        sequence number, so two concurrent submissions would collide. A 200
        with no cycle id means a guard declined it. `usdc_to_usd` is refused
        (400) until a treasury beneficiary configuration exists — see the
        policy endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind: { type: string, enum: [xlm_to_usdc, usdc_to_usd] }
              required: [kind]
      responses:
        '200': { description: Cycle started, or a guard declined it. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: The watcher is mid-tick; retry shortly. }
        '429': { $ref: '#/components/responses/RateLimited' }
  /admin/exchange-reserve/replenishment/{cycle_id}/confirm-fiat:
    post:
      tags: [reserve]
      summary: Confirm a bank credit landed (admin or treasurer; rate limited 5/60s).
      description: |
        The bridge can see USDC leave and the provider's transfer status, but
        never a bank credit — so a completed off-ramp books the fiat into
        `held` as *in transit* and stays there until this call moves it into
        `available`. The one step that cannot be automated honestly.
      parameters:
        - in: path
          name: cycle_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_usd_cents:
                  type: integer
                  format: int64
                  description: Actual cents received, if it differs from the provider's figure.
                external_ref: { type: string }
                note: { type: string }
      responses:
        '200': { description: Confirmed; the float is spendable. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Not awaiting confirmation. }
        '429': { $ref: '#/components/responses/RateLimited' }
  /admin/exchange-reserve/replenishment/{cycle_id}/write-off:
    post:
      tags: [reserve]
      summary: Write off in-transit fiat that never arrived (admin or treasurer).
      description: |
        Writes off real money and is loudly audited; a note is required.
        Without it the USD `held` column stays poisoned and the kind stays
        permanently blocked.
      parameters:
        - in: path
          name: cycle_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
              required: [note]
      responses:
        '200': { description: Written off. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Not awaiting confirmation. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Temporal token (1-hour TTL). Obtain via `POST /token` with a valid
        refresh token. Send as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: impala_session
      description: |
        Browser session cookie (HttpOnly), established by `POST /session/login`
        or `POST /auth/okta` with `cookie_mode: true`. In production the cookie
        is `__Host-`-prefixed (`__Host-impala_session`). Unsafe-method requests
        additionally require the CSRF synchronizer token in the `X-CSRF-Token`
        header.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Token is valid but the resource belongs to another account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "unauthorized" }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    ReplenishPolicyUpdateRequest:
      type: object
      properties:
        enabled: { type: boolean }
        target_days: { type: integer, minimum: 1, maximum: 365 }
        window_days: { type: integer, minimum: 1, maximum: 365 }
        min_need_minor: { type: integer, format: int64, minimum: 0 }
        max_spend_minor:
          type: integer
          format: int64
          minimum: 0
          description: Per-cycle ceiling. 0 means unconfigured — the cycle refuses to run.
        daily_spend_cap_minor:
          type: integer
          format: int64
          minimum: 0
          description: Rolling-24h ceiling. 0 means unconfigured.
        cooldown_secs: { type: integer, minimum: 60, maximum: 604800 }
        min_float_minor:
          type: integer
          format: int64
          minimum: 0
          description: Never spent — must cover the Stellar base reserve plus fees.
        min_price_minor:
          type: integer
          format: int64
          minimum: 0
          description: Recv minor units per whole spend unit. 0 disables the floor.
        max_slippage_bps: { type: integer, minimum: 0, maximum: 5000 }
      required: [enabled, target_days, window_days, min_need_minor, max_spend_minor,
                 daily_spend_cap_minor, cooldown_secs, min_float_minor, min_price_minor,
                 max_slippage_bps]
...